+++ src/icp.c
@@ -1,6 +1,4 @@
 #include <stdio.h>
-#include <stdlib.h>
-#include <stdint.h>
 #include <string.h>
 #include "kd_tree.h"
 #include "vec_math.h"
@@ -24,11 +22,11 @@ void printMat4(float a[4][4]) {
 }
 #endif
 
-void transform(float T[4][4], point4D *points, uint32_t numPts) {
+void transform(float T[4][4], point4D *points, unsigned int numPts) {
 	/*
 	transform - Function to apply current transformation to all points in scan
 	*/
-	uint32_t i;
+	unsigned int i;
 	point4D tmp[numPts];
 	memcpy(tmp, points, numPts*sizeof(point4D));
 	for(i=0; i<numPts; i++) {
@@ -36,11 +34,11 @@ void transform(float T[4][4], point4D *points, uint32_t numPts) {
 	}
 }
 
-static inline void meanBoth(point4D *points1, point4D *points2, float *mean1, float *mean2, uint32_t numPts) {
+static inline void meanBoth(point4D *points1, point4D *points2, float *mean1, float *mean2, unsigned int numPts) {
 	/*
 	meanBoth - Function used to calculate the centroids of scan and model
 	*/
-	uint32_t i;
+	unsigned int i;
 	float pt1[3] = {0.0};
 	float pt2[3] = {0.0};
 	for(i=0; i<numPts; i++) {
@@ -63,7 +61,7 @@ static inline void meanBoth(point4D *points1, point4D *points2, float *mean1, fl
 
 
 
-void calcTransform(point4D *scan, point4D *model, float T[4][4], uint32_t numPts) {
+void calcTransform(point4D *scan, point4D *model, float T[4][4], unsigned int numPts) {
 	/*
 	calcTransform - Function to calculate the current optimal transformation to minimize distance
 	Inputs:
@@ -72,7 +70,7 @@ void calcTransform(point4D *scan, point4D *model, float T[4][4], uint32_t numPts
 		T 		- Current transformation matrix
 		numPts	- Number of points in scan/model
 	*/
-	uint32_t i;
+	unsigned int i;
 	float W[3][3] = {0.0};
 	point3D centScan, centModel;
 	point4D tempScan[numPts], tempModel[numPts];
@@ -106,7 +104,7 @@ void calcTransform(point4D *scan, point4D *model, float T[4][4], uint32_t numPts
 	T[2][3] = t[2];
 }
 
-float icp(point4D *scan, node *root, float T[4][4], uint32_t numPts, uint8_t iterations) {
+float icp(point4D *scan, node *root, float T[4][4], unsigned int numPts, unsigned int iterations) {
 	/*
 	icp - Function to control iterative closest point calculation
 	Inputs:
@@ -115,8 +113,8 @@ float icp(point4D *scan, node *root, float T[4][4], uint32_t numPts, uint8_t ite
 		T	  		- Initial transformation matrix
 		numPts	- Number of points in the flash lidar point cloud
 	*/
-	uint8_t i;
-	uint32_t numKeep;
+	unsigned int i;
+	unsigned int numKeep;
 	point4D initState[numPts], closestPts[numPts];
 	point4D modelPts[numPts], scanPts[numPts];
 	memcpy(initState, scan, numPts*sizeof(point4D));
